requirements.txt
keyboard 
pyautogui 
pytesseract 
pillow 
mistralai